add bittorrent and tor to be enginerequest types
